add client caching by config, extract client and logging packages
It doesn't make sense for each provider (sdk v2 + framework) to initialize its own client because it has to go through ping and service discovery requests. This may also slightly speed up testing since each test begins by initializing a new client.
