+++ internal/provider/plugin_provider.go
@@ -11,6 +11,7 @@ import (
 	tfe "github.com/hashicorp/go-tfe"
 	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
+	"github.com/hashicorp/terraform-provider-tfe/internal/client"
 )
 
 type pluginProviderServer struct {
@@ -71,7 +72,7 @@ func (p *pluginProviderServer) ConfigureProvider(ctx context.Context, req *tfpro
 		return resp, nil
 	}
 
-	client, err := getClient(meta.hostname, meta.token, meta.sslSkipVerify)
+	client, err := client.GetClient(meta.hostname, meta.token, meta.sslSkipVerify)
 	if err != nil {
 		resp.Diagnostics = append(resp.Diagnostics, &tfprotov5.Diagnostic{
 			Severity: tfprotov5.DiagnosticSeverityError,
